test: add property-based tests for romeo asset contract
Introduce fast-check, the property-based testing lib of
choice in TypeScript.

One of the features of fast-check is that it can create
extreme/larger values with higher probability for first
runs and lower ones as you go in the runs.

Contrary to the solutions adopted by QuickCheck (v1/v2)
and many others where the first runs must only generate
small values.

Doing that may miss edge cases, as some bugs may appear
for long arrays with only small values.

Find out more about the concept in the discussion I had
with Nicolas Dubien the author of fast-check:
https://github.com/dubzzz/fast-check/discussions/2978